refactor(catalog): extract pages extensions file
